requirements.txt
haystack-ai
google-genai-haystack
datasets
pandas
sentence-transformers
python-dotenv
streamlit
